requirements.txt
streamlit==1.38.0
pandas==2.2.3
joblib==1.4.2
cachetools==5.5.0
python-dotenv==1.0.1
langchain==0.3.1
langchain-community==0.3.1
scikit-learn==1.6.1